backend/requirements.txt
alabaster==0.7.12
amqp==5.0.6
argh==0.26.2
attrs==20.3.0
Babel==2.9.0
blueprint==3.4.2
boto3
celery==5.1.1
certifi==2020.11.8
expiringdict==1.2.1
flake8==3.8.4
flake8-polyfill==1.0.2
flasgger==0.9.5
Flask==1.1.2
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
html2text==2020.1.16
humanize==3.9.0
idna==2.10
imagesize==1.2.0
isort==5.6.4
itsdangerous==1.1.0
jenkinsapi==0.3.11
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
packaging==20.4
pathtools==0.1.2
pdfkit==0.6.1
pep257==0.7.0
pep8-naming==0.10.0
pkcs7==0.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pycryptodome==3.9.9
pycurl==7.43.0.6; sys_platform != "win32"
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==7.1.2
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requirements-detector==0.7
s3transfer==0.3.3
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.2
tornado==6.1
ua-parser==0.10.0
unicodecsv==0.14.1
urllib3==1.26.2
vine==5.0.0
watchdog==0.10.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.7
uwsgitop==0.11
wheel
redis==3.5.3
numpy==1.21.2
scipy==1.7.3
pyspark==3.1.2
pyparsing==2.4.7
mysql-connector-python
Flask-Cors==3.0.10
Flask-Login==0.6.2
coverage==7.2.5
stripe==5.4.0
pandas
openpyxl
